Added Actors
Discovered that while the factories don't show the objects inside the context menu they were exposing them to elsewhere. I also realised that PlanetSystem and PlanetObject had to be actors in order to be placed and used in the level. This also made them appear as blueprint classes in the context menu which is progress.